[MOV] mail: split some functions and parameters (portal)
Move functions and parameters related to `followers, recipients, suggested
recipients and activities`
